requirements.txt
pysnmp
requests
redis
PyYAML
